[feat] `ManualNumberList` 수동으로 입력받는 NumberList Strategy 추가
